Extra_challenge/mastadon_ETL/etl_service/requirements.txt
psycopg2-binary
requests
vaderSentiment
python-dotenv